+++ FastLED_NeoMatrix.h
@@ -45,11 +45,11 @@ class FastLED_NeoMatrix : public Framebuffer_GFX {
 
 public:
   // Constructor for single matrix:
-  FastLED_NeoMatrix(CRGB *, uint8_t w, uint8_t h, 
+  FastLED_NeoMatrix(CRGB *, uint16_t w, uint16_t h, 
     uint8_t matrixType = NEO_MATRIX_TOP + NEO_MATRIX_LEFT + NEO_MATRIX_ROWS);
 
   // Constructor for tiled matrices:
-  FastLED_NeoMatrix(CRGB *, uint8_t matrixW, uint8_t matrixH, 
+  FastLED_NeoMatrix(CRGB *, uint16_t matrixW, uint16_t matrixH, 
     uint8_t tX, uint8_t tY,
     uint8_t matrixType = NEO_MATRIX_TOP + NEO_MATRIX_LEFT + NEO_MATRIX_ROWS +
                          NEO_TILE_TOP + NEO_TILE_LEFT + NEO_TILE_ROWS);
